extract PageViewHolder from Page model
