+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/model/Header.java
@@ -11,7 +11,6 @@ import android.widget.TextView;
 import org.ccci.gto.android.common.util.XmlPullParserUtils;
 import org.cru.godtools.tract.R;
 import org.cru.godtools.tract.R2;
-import org.cru.godtools.tract.model.Page.PageViewHolder;
 import org.xmlpull.v1.XmlPullParser;
 import org.xmlpull.v1.XmlPullParserException;
